A Basic Team Management App using Django
This is a simple team-member management application that allows the user to view, edit, add, and delete team members. The app consists of 3 pages:

List Page: This page shows a list of all team members.
Edit Page: This page lets the user edit an existing team member
Add page: This page lets user to add a new team member